// app/Models/cakerequests.ts
export type CakeRequest = {
    id?: number;
    messageOnCake?: string;
    deliveryDate?: Date | string;
    eggEggles?: string;
    email?: string;
    phone?: number;
    pincode?: number;
    address?: string;
    cakeName?: string;
    quantity?: number;
    price?: number;
    totalPrice?: number;
}